requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
idna==3.9
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
